// frontend/src/types/components/ContractType.ts
export interface LeaseContractType {
  contractDate: string // 계약날짜
  deposit: {
    _hex: string
  } // 보증금
  landlordDID: string // 임대인 지갑
  leasePeriod: number // 임대긴간
  propertyDID: string // 부동산 did 주소
  tenantDID: string // 임차인 did 주소
  terms: string // 특약사항
}